requirements-langchain.txt
google-search-results
openai
langchain
python-dotenv
streamlit
faiss-cpu
tiktoken